src/kit-app-template/source/mqtt_sim/requirements.txt
pandas
paho-mqtt
python-dotenv
Flask